refactor(splits): moving simulation operations to assets module
